got the hero x,y debug printing working right. had to specify a width. before it'd wind up printing something like X: 400 when the hero was at 4, the 00 left over from a printout back when it was at some place >= 100.

+++ source/arm9/include/hero.h
@@ -90,8 +90,8 @@ public:
 		{
 			// ansi escape sequence to set print co-ordinates
 			// /x1b[line;columnH
-			iprintf("\x1b[1;24HX: %d\n", (int) position.x);
-			iprintf("\x1b[2;24HY: %d\n", (int) position.y);
+			iprintf("\x1b[1;24HX: %5ld\n", (long int) position.x);
+			iprintf("\x1b[2;24HY: %5ld\n", (long int) position.y);
 		}
 		++numCalls;
